feat: integrate QA Playground project, implement global toast notification system, and resolve resume download asset sync issues
