+++ CHANGELOG.md
@@ -6,13 +6,55 @@ Maintained date-wise in descending order (latest first).
 ---
 
 ## [Unreleased]
+
 > Changes that are in progress or planned but not yet committed.
+
+---
+
+## 2026-04-14
+
+### Added
+
+- **QA Automation Sandbox** — Integrated the newly deployed "QA Playground" testing sandbox into the portfolio as a top-level featured project (`profile.js`).
+- **Global Header CTA** — Added a standalone, glowing "Try QA Playground" CTA button (`animate-ping`) into the main Header navigation, perfectly load-balanced using flexbox layout spacing.
+
+### Fixed
+
+- **Resume Download Hook** — Fixed an exact case-sensitivity tracking bug for the resume PDF (`/Lakshya...` vs `/lakshya...`) in `profile.js` that was failing the Vercel strict `HEAD` request pre-flight checks, which triggered the "server timeout" error toast.
+
+---
+
+## 2026-04-13
+
+### Added
+
+- **Smart Resume Discovery** — Created `scripts/sync-resume.js` to automatically identify the latest resume PDF in `/public` using keyword matching and timestamps.
+- **Creative Toast System** — Implemented a global `ToastProvider` and `useToast` hook for dynamic, "QA-persona" themed notifications.
+- **Custom React Hooks** — Added `useResumeDownload.js` to handle secure file verification and creative feedback logic.
+- **Modern UI Components** — Created a sleek, animated `Toast.jsx` component using `framer-motion` and glassmorphism.
+
+### Changed
+
+- **Professional Profile** — Refactored `src/data/profile.js` to accurately reflect recent experience at TuningBill Services (Jaipur), updated tech stack (Playwright, Appium, Agentic AI), and reorganized professional projects.
+- **EmailJS Security** — Moved all EmailJS credentials from hardcoded constants in `Contact.jsx` to environment variables (`VITE_EMAILJS_*`).
+- **Dark Mode** — Upgraded dark mode implementation to use Tailwind v4 `@custom-variant` for better component-level consistency.
+- **Build Workflow** — Integrated `sync-resume.js` into `package.json` dev and build hooks to automate asset management.
+- **Contact Form Feedback** — Replaced legacy inline status text with site-wide creative toasts for form submissions and email/phone copy actions.
+
+### Fixed
+
+- **Resume Download** — Resolved a 404/400 error on production (Vercel) caused by case-sensitivity mismatches and missing build-time asset syncing.
+- **SPA Fallback Handled** — Hardened the `useResumeDownload` hook to check the `Content-Type` header, preventing Vite and Vercel from falsely serving a 6KB `index.html` fallback as a successful PDF download.
+- **Contact Form Reliability** — Fixed a background syntax error in `Contact.jsx` that was preventing successful form submissions.
+- **Asset Integrity** — Renamed default resume to a professional standard `Lakshya_Sharma_QA_Resume.pdf` and ensured it's tracked by Git in the `/public` folder.
+- **Constants Extraction** — Extracted all "QA Humor" toast messages into a dedicated `messages.js` constants file to resolve React Fast Refresh warnings.
 
 ---
 
 ## 2026-03-14
 
 ### Added
+
 - **CHANGELOG.md** — Created this changelog file to track all project changes going forward
 - **`@tailwindcss/postcss`** — Installed as required PostCSS plugin for Tailwind v4
 - **`.prettierrc`** — Added Prettier configuration (semi, singleQuote, tabWidth, trailingComma, printWidth, etc.)
@@ -20,6 +62,7 @@ Maintained date-wise in descending order (latest first).
 - **`public/lakshyaSharmaQaAutomationEngineerResume(1).pdf`** — Uploaded updated resume PDF copy
 
 ### Changed
+
 - **`package.json`** — Renamed project from `app` to `lakshyasharmaqa-portfolio`, bumped version to `1.0.0`, updated Node engine to `24.x`, updated Volta node pin to `24.13.1`, added `@tailwindcss/postcss` to dependencies, added `lint:fix`, `format`, and `format:check` scripts, added `prettier` as devDependency, moved `framer-motion`, `react-intersection-observer`, `react-typed`, `recharts` to `dependencies` (from devDependencies), upgraded `tailwindcss` from `v3` → `v4`, upgraded `vite` from `v7` → `v6`, upgraded `react` and `react-dom` to `v19.2.4`
 - **`package-lock.json`** — Updated to reflect new `@tailwindcss/postcss` package and all dependency upgrades
 - **`postcss.config.js`** — Replaced `tailwindcss: {}` with `'@tailwindcss/postcss': {}` — required breaking change in Tailwind v4
@@ -33,6 +76,7 @@ Maintained date-wise in descending order (latest first).
 - **`README.md`** — Full rewrite: cleaner structure, removed bloated sections, added tech stack table, improved local setup steps, added detailed Vercel deployment guide with auto-deploy explanation; added `Deploy via Vercel CLI` section with all CLI commands: `vercel login`, `vercel`, `vercel --prod`, `vercel ls`, `vercel inspect`
 
 ### Fixed
+
 - **`package.json`** — Removed UTF-8 BOM character (`EF BB BF`) from start of file that was causing Vercel deployment to fail with `Unexpected token` JSON parse error
 - **`.prettierrc`** — Removed UTF-8 BOM character (same issue)
 - **`.prettierignore`** — Removed UTF-8 BOM character (same issue)
@@ -43,6 +87,7 @@ Maintained date-wise in descending order (latest first).
 - **Timeline UI** — Replaced visually distracting solid green left border line with cleaner dashed style
 
 ### Removed
+
 - **`.emergent/emergent.yml`** — Deleted auto-generated Emergent agent config file (not needed)
 - **`.gitconfig`** — Deleted auto-generated Emergent agent git config (not needed)
 
@@ -53,6 +98,7 @@ Maintained date-wise in descending order (latest first).
 ## 2025-10-01
 
 ### Added
+
 - Uploaded `ISTQB Result.png` certification image to the project
 
 ---
@@ -60,6 +106,7 @@ Maintained date-wise in descending order (latest first).
 ## 2025-09-23
 
 ### Fixed
+
 - Replaced profile image (`profile-portfolio.png`) with corrected version
 
 ---
@@ -67,10 +114,12 @@ Maintained date-wise in descending order (latest first).
 ## 2025-09-22
 
 ### Added
+
 - Added new profile image (`profile-portfolio.png`) to assets
 - Added `package-lock.json` to the repository
 
 ### Changed
+
 - Updated `.gitignore` for UAT branch testing
 
 ---
@@ -78,6 +127,7 @@ Maintained date-wise in descending order (latest first).
 ## 2025-09-21
 
 ### Changed
+
 - **Header.jsx** — Updated name display on top-left of navigation bar
 - **Hero.jsx** — Updated hero section with profile image and layout improvements
 - **Resume.jsx** — Resume "Last Updated" now dynamically syncs to current month and year (no manual updates needed)
@@ -85,6 +135,7 @@ Maintained date-wise in descending order (latest first).
 - **Assets** — Uploaded updated image assets
 
 ### Removed
+
 - Deleted unused `public/vite.svg` default Vite asset
 
 ---
@@ -92,10 +143,12 @@ Maintained date-wise in descending order (latest first).
 ## 2025-09-20
 
 ### Added
+
 - Initial project setup — React 19 + Vite + Tailwind CSS portfolio
 - Uploaded `resume.pdf` to `/public` folder
 
 ### Changed
+
 - **Contact.jsx** — Integrated EmailJS with actual service ID, template ID, and API keys for functional contact form
 - **About.jsx** — Updated section content and text copy
 - **profile.js** — Updated personal information, skills, experience, and project data
@@ -103,6 +156,7 @@ Maintained date-wise in descending order (latest first).
 - **sitemap.xml** — Updated sitemap with correct URLs
 
 ### Fixed
+
 - Renamed resume file from `resume.pdf.pdf` → `resume.pdf` (fixed double extension issue)
 
 ---
