--- src/context/ToastContext.jsx
@@ -1,34 +1,0 @@
-import React, { createContext, useContext, useState, useCallback } from 'react'
-
-const ToastContext = createContext(null)
-
-export const ToastProvider = ({ children }) => {
-  const [toasts, setToasts] = useState([])
-
-  const addToast = useCallback((message, type = 'success', duration = 5000) => {
-    const id = Math.random().toString(36).substring(2, 9)
-    setToasts(prev => [...prev, { id, message, type }])
-
-    setTimeout(() => {
-      removeToast(id)
-    }, duration)
-  }, [])
-
-  const removeToast = useCallback(id => {
-    setToasts(prev => prev.filter(toast => toast.id !== id))
-  }, [])
-
-  return (
-    <ToastContext.Provider value={{ addToast, removeToast, toasts }}>
-      {children}
-    </ToastContext.Provider>
-  )
-}
-
-export const useToast = () => {
-  const context = useContext(ToastContext)
-  if (!context) {
-    throw new Error('useToast must be used within a ToastProvider')
-  }
-  return context
-}
