+++ vite.config.js
@@ -5,6 +5,10 @@ import react from '@vitejs/plugin-react'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  server: {
+    port: 5252,
+    host: true,
+  },
   test: {
     globals: true,
     environment: 'jsdom',
